feat: refactor backend code
