+++ app.py
@@ -4,7 +4,6 @@ from api_url import apibp
 app = Flask(__name__)
 app.config["JSON_AS_ASCII"] = False
 app.config["TEMPLATES_AUTO_RELOAD"] = True
-app.config["JSONIFY_MIMETYPE"] = "application/json; charset=utf-8"
 
 
 # Pages
